fix: 예약 테이블 컬럼 변경으로 CRUD 수정, ID 타입을 int에서 long으로 수정

+++ bible/src/main/java/com/library/bible/reservation/service/IReservationService.java
@@ -7,8 +7,8 @@ import com.library.bible.reservation.model.Reservation;
 public interface IReservationService {
 	List<Reservation> selectAllReserv();
 	
-	Reservation selectReserv(int reservId);
+	Reservation selectReserv(long reservId);
 	void insertReserv(Reservation reservation);
 	void updateReserv(Reservation reservation);
-	int deleteReserv(int reservId);
+	int deleteReserv(long reservId);
 }
